+++ saleor/plugins/manager.py
@@ -76,7 +76,6 @@ if TYPE_CHECKING:
         ProductType,
         ProductVariant,
     )
-    from ..shipping.interface import ShippingMethodData
     from ..shipping.models import ShippingMethod, ShippingZone
     from ..site.models import SiteSettings
     from ..tax.models import TaxClass
@@ -2575,36 +2574,6 @@ class PluginsManager(PaymentInterface):
                 )
             )
         return gateways
-
-    def list_shipping_methods_for_checkout(
-        self,
-        checkout: "Checkout",
-        built_in_shipping_methods: list["ShippingMethodData"],
-        channel_slug: str | None = None,
-        active_only: bool = True,
-    ) -> list["ShippingMethodData"]:
-        channel_slug = channel_slug if channel_slug else checkout.channel.slug
-        plugins = self.get_plugins(channel_slug=channel_slug, active_only=active_only)
-        shipping_plugins = [
-            plugin
-            for plugin in plugins
-            if hasattr(plugin, "get_shipping_methods_for_checkout")
-        ]
-
-        shipping_methods = []
-        for plugin in shipping_plugins:
-            shipping_methods.extend(
-                # https://github.com/python/mypy/issues/9975
-                getattr(plugin, "get_shipping_methods_for_checkout")(
-                    checkout, built_in_shipping_methods, None
-                )
-            )
-        return list(
-            filter(
-                lambda method: method.price.currency == checkout.currency,
-                shipping_methods,
-            )
-        )
 
     def list_external_authentications(self, active_only: bool = True) -> list[dict]:
         auth_basic_method = "external_obtain_access_tokens"
